- Reemplaza BUFFER_SALIDA único por BUFFER_SALIDA_A/B (ping-pong) - DSP escribe siempre en el buffer inactivo (no usado por el DAC) - DMA del DAC arranca recién cuando el primer buffer está lleno,   evitando condición de carrera en el inicio - DAC_BUFFER_ACTIVO inicializado en 1U para que el primer ciclo   escriba en BUFFER_SALIDA_A y el DMA arranque desde ahí - ConfDMA_BUFFERSALIDA_DAC() movido fuera de main(), se llama   desde Procesamiento_Senales() en la primera iteración - Corregido buf_lectura: apuntaba a BUFFER_SALIDA en vez de BUFFERA/B - Corregida lógica de buf_escritura invertida respecto a DAC_BUFFER_ACTIVO - InicializarFiltros() agregado en dsp.c con coeficientes Butterworth   biquad RBJ (Fs=44100Hz): LP 250Hz / BP 250-4000Hz / HP 4000Hz - InicializarFiltros() se llama en main() antes de cualquier periférico - Correcion del tipo de dato enviado en funciones de UART

+++ src/dac.c
@@ -20,5 +20,6 @@ void ConfDAC(void)
 
     /* Timeout = DAC_DMA_TIMEOUT ticks de PCLK_DAC
      * Determina la frecuencia de muestreo de salida. */
-    DAC_SetDMATimeOut(DAC_DMA_TIMEOUT);
+    DAC_SetDMATimeOut((uint16_t)DAC_DMA_TIMEOUT);
 }
+
